removing dtos for another project, refactoring
