+++ api-products/pom.xml
@@ -49,6 +49,11 @@
 				<groupId>org.postgresql</groupId>
 				<artifactId>postgresql</artifactId>
 		</dependency>
+		<dependency>
+				<groupId>com.microservices</groupId>
+				<artifactId>shopping-client</artifactId>
+				<version>0.0.1-SNAPSHOT</version>
+		</dependency>
 		</dependencies>
 
 	<build>
