+++ api-shop/src/main/java/com/microservices/apishop/service/ShopService.java
@@ -7,9 +7,11 @@ import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
 
-import com.microservices.apishop.dto.ShopDto;
+import com.microservices.apishop.dto.DtoConverter;
+//import com.microservices.apishop.dto.ShopDto;
 import com.microservices.apishop.model.Shop;
 import com.microservices.apishop.repository.ShopRepository;
+import com.microservices.shoppingclient.dto.ShopDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -20,24 +22,25 @@ public class ShopService {
 
 	public List<ShopDto> getAll() {
 		List<Shop> shops = shopRepository.findAll();
-		return shops.stream().map(ShopDto::convert).collect(Collectors.toList());
+//		return shops.stream().map(ShopDto::convert).collect(Collectors.toList());
+		return shops.stream().map(DtoConverter::convert).collect(Collectors.toList());
 	}
 
 	public List<ShopDto> getByUser(String userIdentifier) {
 		List<Shop> shops = shopRepository.findAllByUserIdentifier(userIdentifier);
 
-		return shops.stream().map(ShopDto::convert).collect(Collectors.toList());
+		return shops.stream().map(DtoConverter::convert).collect(Collectors.toList());
 	}
 
 	public List<ShopDto> getByDate(ShopDto shopDto) {
 		List<Shop> shops = shopRepository.findAllByDateGreaterThan(shopDto.getDate());
-		return shops.stream().map(ShopDto::convert).collect(Collectors.toList());
+		return shops.stream().map(DtoConverter::convert).collect(Collectors.toList());
 	}
 
 	public ShopDto findById(long ProductId) {
 		Optional<Shop> shop = shopRepository.findById(ProductId);
 		if (shop.isPresent()) {
-			return ShopDto.convert(shop.get());
+			return DtoConverter.convert(shop.get());
 		}
 		return null;
 	}
@@ -47,6 +50,6 @@ public class ShopService {
 		Shop shop = Shop.convert(shopDto);
 		shop.setDate(LocalDateTime.now());
 		shop = shopRepository.save(shop);
-		return ShopDto.convert(shop);
+		return DtoConverter.convert(shop);
 	}
 }
