+++ api-user/src/main/java/com/microservices/apiuser/service/UserService.java
@@ -8,9 +8,10 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import com.microservices.apiuser.dto.UserDto;
+import com.microservices.apiuser.dto.DtoConverter;
 import com.microservices.apiuser.model.User;
 import com.microservices.apiuser.repository.UserRepository;
+import com.microservices.shoppingclient.dto.UserDto;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,43 +22,43 @@ public class UserService {
 	
 	public List<UserDto> getAll(){
 		List<User> users = userRepository.findAll();
-		return users.stream().map(UserDto::convert)
+		return users.stream().map(DtoConverter::convert)
 				.collect(Collectors.toList());
 	}
 	
 	public UserDto findById(long userId) {
 		User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("user not found"));
-		return UserDto.convert(user);
+		return DtoConverter.convert(user);
 	}
 	
 	public UserDto save(UserDto userDto) {
 		userDto.setDataCadastro(LocalDateTime.now());
 		User user = userRepository.save(User.convert(userDto));
-		return UserDto.convert(user);
+		return DtoConverter.convert(user);
 	}
 	
 	public UserDto delete(long userId) {
 		User user = userRepository.findById(userId).orElseThrow(()-> new RuntimeException());
 		userRepository.delete(user);
-		return UserDto.convert(user);
+		return DtoConverter.convert(user);
 	}
 	
 	public UserDto findByCpf(String cpf) {
 		User user = userRepository.findByCpf(cpf);
 		if(user != null) {
-			return UserDto.convert(user);
+			return DtoConverter.convert(user);
 		}
 		return null;
 	}
 	
 	public List<UserDto> queryByNome(String nome){
 		List<User> users = userRepository.queryByNomeLike(nome);
-		return users.stream().map(UserDto::convert).collect(Collectors.toList());		
+		return users.stream().map(DtoConverter::convert).collect(Collectors.toList());		
 	}
 	
 	public List<UserDto> queryByNomeContaining(String nome){
 		List<User> users = userRepository.queryByNomeContaining(nome);
-		return users.stream().map(UserDto::convert).collect(Collectors.toList());		
+		return users.stream().map(DtoConverter::convert).collect(Collectors.toList());		
 	}
 	
 	public UserDto editUser(Long userId, UserDto userDto) {
@@ -72,13 +73,13 @@ public class UserService {
 			user.setEndereco(userDto.getEndereco());
 		}
 		user = userRepository.save(user);
-		return UserDto.convert(user);
+		return DtoConverter.convert(user);
 	
 	}
 	
 	//com paginação
 	public Page<UserDto> getAllPage(Pageable page){
 		Page<User> users = userRepository.findAll(page);
-		return users.map(UserDto::convert);
+		return users.map(DtoConverter::convert);
 	}
 }
